+++ src/textual/message_pump.py
@@ -725,6 +725,8 @@ class MessagePump(metaclass=_MessagePumpMeta):
                                 continue
                             for attribute, selector in selectors.items():
                                 node = getattr(message, attribute)
+                                if node is None:
+                                    break
                                 if not isinstance(node, Widget):
                                     raise OnNoWidget(
                                         f"on decorator can't match against {attribute!r} as it is not a widget."